pyproject.toml
[project]
name = "budget-tracker"
version = "0.1.0"
description = "A personal budget tracking application built with FastAPI and SQLite."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt<4.1",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
